Only add cache layer for #[cfg(not(windows))] and fix object store upgrade
