+++ ballista/core/src/lib.rs
@@ -22,12 +22,14 @@ pub fn print_version() {
     println!("Ballista version: {BALLISTA_VERSION}")
 }
 
+#[cfg(not(windows))]
 pub mod cache_layer;
 pub mod client;
 pub mod config;
 pub mod error;
 pub mod event_loop;
 pub mod execution_plans;
+pub mod object_store_registry;
 /// some plugins
 pub mod plugin;
 pub mod utils;
