+++ ballista/executor/src/executor_process.rs
@@ -40,11 +40,15 @@ use uuid::Uuid;
 use datafusion::execution::runtime_env::{RuntimeConfig, RuntimeEnv};
 use datafusion_proto::protobuf::{LogicalPlanNode, PhysicalPlanNode};
 
-use ballista_core::cache_layer::medium::local_disk::LocalDiskMedium;
-use ballista_core::cache_layer::policy::file::FileCacheLayer;
-use ballista_core::cache_layer::CacheLayer;
+#[cfg(not(windows))]
+use ballista_core::cache_layer::{
+    medium::local_disk::LocalDiskMedium, policy::file::FileCacheLayer, CacheLayer,
+};
 use ballista_core::config::{DataCachePolicy, LogRotationPolicy, TaskSchedulingPolicy};
 use ballista_core::error::BallistaError;
+#[cfg(not(windows))]
+use ballista_core::object_store_registry::cache::with_cache_layer;
+use ballista_core::object_store_registry::with_object_store_registry;
 use ballista_core::serde::protobuf::executor_resource::Resource;
 use ballista_core::serde::protobuf::executor_status::Status;
 use ballista_core::serde::protobuf::{
@@ -54,8 +58,7 @@ use ballista_core::serde::protobuf::{
 };
 use ballista_core::serde::BallistaCodec;
 use ballista_core::utils::{
-    create_grpc_client_connection, create_grpc_server, get_time_before, with_cache_layer,
-    with_object_store_registry,
+    create_grpc_client_connection, create_grpc_server, get_time_before,
 };
 use ballista_core::BALLISTA_VERSION;
 
@@ -182,29 +185,35 @@ pub async fn start_executor_process(opt: Arc<ExecutorProcessConfig>) -> Result<(
         }),
     };
 
-    let cache_dir = opt.cache_dir.clone();
-    let cache_capacity = opt.cache_capacity;
-    let cache_io_concurrency = opt.cache_io_concurrency;
-    let cache_layer: Option<CacheLayer> =
-        opt.data_cache_policy
-            .map(|data_cache_policy| match data_cache_policy {
-                DataCachePolicy::LocalDiskFile => {
-                    let cache_dir = cache_dir.unwrap();
-                    let cache_layer = FileCacheLayer::new(
-                        cache_capacity as usize,
-                        cache_io_concurrency,
-                        LocalDiskMedium::new(cache_dir),
-                    );
-                    CacheLayer::LocalDiskFile(Arc::new(cache_layer))
-                }
-            });
-
     let config = RuntimeConfig::new().with_temp_file_path(work_dir.clone());
-    let config = if let Some(cache_layer) = cache_layer {
-        with_cache_layer(config, cache_layer)
-    } else {
-        with_object_store_registry(config)
+    // Set the object store registry
+    #[cfg(not(windows))]
+    let config = {
+        let cache_dir = opt.cache_dir.clone();
+        let cache_capacity = opt.cache_capacity;
+        let cache_io_concurrency = opt.cache_io_concurrency;
+        let cache_layer =
+            opt.data_cache_policy
+                .map(|data_cache_policy| match data_cache_policy {
+                    DataCachePolicy::LocalDiskFile => {
+                        let cache_dir = cache_dir.unwrap();
+                        let cache_layer = FileCacheLayer::new(
+                            cache_capacity as usize,
+                            cache_io_concurrency,
+                            LocalDiskMedium::new(cache_dir),
+                        );
+                        CacheLayer::LocalDiskFile(Arc::new(cache_layer))
+                    }
+                });
+        if let Some(cache_layer) = cache_layer {
+            with_cache_layer(config, cache_layer)
+        } else {
+            with_object_store_registry(config)
+        }
     };
+    #[cfg(windows)]
+    let config = with_object_store_registry(config);
+
     let runtime = Arc::new(RuntimeEnv::new(config).map_err(|_| {
         BallistaError::Internal("Failed to init Executor RuntimeEnv".to_owned())
     })?);
