+++ ballista/executor/src/standalone.rs
@@ -18,13 +18,14 @@
 use crate::metrics::LoggingMetricsCollector;
 use crate::{execution_loop, executor::Executor, flight_service::BallistaFlightService};
 use arrow_flight::flight_service_server::FlightServiceServer;
-use ballista_core::serde::scheduler::ExecutorSpecification;
-use ballista_core::serde::BallistaCodec;
-use ballista_core::utils::{create_grpc_server, with_object_store_registry};
 use ballista_core::{
     error::Result,
+    object_store_registry::with_object_store_registry,
     serde::protobuf::executor_registration::OptionalHost,
     serde::protobuf::{scheduler_grpc_client::SchedulerGrpcClient, ExecutorRegistration},
+    serde::scheduler::ExecutorSpecification,
+    serde::BallistaCodec,
+    utils::create_grpc_server,
     BALLISTA_VERSION,
 };
 use datafusion::execution::runtime_env::{RuntimeConfig, RuntimeEnv};
